feat(search): implement JackettCardMatcher and JackettResultBuilder for torrent search functionality
- Introduced `JackettCardMatcher` for processing search queries and matching torrents based on various criteria, including title, year, and category.
- Added `JackettResultBuilder` to manage the addition and merging of torrent results, ensuring efficient handling of duplicates and metadata.
- Refactored `JackettSearchService` to utilize the new matcher and result builder, streamlining the search process and improving code organization.
- Enhanced regex handling for query parsing and category identification, allowing for more accurate search results.
